requirements.txt
# Core dependencies for job scraper
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
webdriver-manager>=4.0.1
python-dotenv>=1.0.0
lxml>=4.9.3

# Optional dependencies for advanced features
cloudscraper>=1.2.71
fake-useragent>=1.4.0
urllib3>=2.0.0

# Utilities
tqdm>=4.66.0
